requirements.txt
streamlit
duckduckgo_search
python-docx
pandas